+++ Bugsnag/Client/BugsnagClient.m
@@ -284,6 +284,8 @@ BSG_OBJC_DIRECT_MEMBERS
 #endif
                  object:nil];
 
+    self.ready = YES;
+
     id<BugsnagPlugin> reactNativePlugin = [NSClassFromString(@"BugsnagReactNativePlugin") new];
     if (reactNativePlugin) {
         [self.configuration.plugins addObject:reactNativePlugin];